+++ examples/21_ide_project/top.v
@@ -1,0 +1,3 @@
+module top(input clk, output led);
+    assign led = clk;
+endmodule